+++ test/integration/app.js
@@ -1,5 +1,6 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
+const cheerio = require('cheerio');
 
 const constants = require('../../app/lib/constants');
 const server = require('../../server');
@@ -26,5 +27,14 @@ describe('An unknown page', () => {
       expect(err).to.have.status(404);
       expect(err.response).to.be.html;
     }
+  });
+});
+
+describe('meta tags', () => {
+  it('should instruct Webtrends to anonymise IP Addresses', async () => {
+    const res = await chai.request(server).get('/');
+    const $ = cheerio.load(res.text);
+
+    expect($('meta[name="DCS.dcsipa"]').prop('content')).to.equal('1');
   });
 });